pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodal-final"
version = "0.1.0"
description = "Publication-oriented multimodal pneumonia detection on MIMIC-CXR + MIMIC-IV/MIMIC-IV-ED"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "pyarrow>=14.0",
    "streamlit>=1.35",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "torch>=2.2",
    "torchvision>=0.17",
    "pillow>=10.0",
    "tqdm>=4.66",
    "joblib>=1.3",
    "pyyaml>=6.0",
    "opencv-python>=4.8",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.black]
line-length = 100
target-version = ["py310"]